put copy logic into move dialog
